rm mk

+++ src/Decomposition/tests/ORBDecompositionStrategy_unit_test.hpp
@@ -21,7 +21,7 @@ const size_t type = 0;
 
 // Type of the vector containing particles
 constexpr unsigned int SPACE_N_DIM = 3;
-using domain_type = double;
+using domain_type = float;
 
 // n of points sampled
 #define N_POINTS 1024
@@ -65,16 +65,21 @@ void ORBDecomposition_non_periodic_test(const unsigned int nProcs) {
 
   //////////////////////////////////////////////////////////////////// decompose
   ORB<SPACE_N_DIM, domain_type> orb(box, 16, vp);
-
   // dec.decompose(vp);
 
   /////////////////////////////////////////////////////////////////// distribute
   // dist.distribute(dec.inner().getGraph());
 
   //////////////////////////////////////////////////////////////////////// merge
-  // no need, we're NOT using sub(sub)domains
+  // no need (we're NOT using sub(sub)domains)
 
-  // todo asserts
+  // ASSERT: i-th processing unit has i-th vertex
+  // auto myRank = vcl.rank();
+  // auto myLeaf = dec.inner().getGraph().template vertex_p<nm_v_proc_id>(myRank);
+  // printMe(vcl);
+  // std::cout << "assert " << myRank << " == " << myLeaf << std::endl;
+
+  // todo more asserts
 }
 
 #endif // SRC_DECOMPOSITION_ORB_DECOMPOSITIONSTRATEGY_UNIT_TEST_HPP
